+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.4.0
 
  - Update PerfDatum field names to make them easier to use.
+ - Remove `UnknownUOM` constructor (deprecated in 0.3.2).
 
 ## v0.3.2